@@ -1,28 +1,24 @@
-from typing import List
-
 from nonebot import get_driver
 from pydantic import BaseSettings, validator
 
 from src.utils.helpers import groupidtostr, strtogroupid
 from src.utils.plugin import PluginData
 
-DATA = PluginData('morning', config=True)
+DATA = PluginData('morning')
 
 
 class Config(BaseSettings):
-    morning_hour: int = int(DATA.get_config('morning', 'hour', fallback='7'))
-    morning_minute: int = int(
-        DATA.get_config('morning', 'minute', fallback='30'))
-    morning_second: int = int(
-        DATA.get_config('morning', 'second', fallback='0'))
+    morning_hour: int = DATA.config.getint('morning', 'hour', fallback=7)
+    morning_minute: int = DATA.config.getint('morning', 'minute', fallback=30)
+    morning_second: int = DATA.config.getint('morning', 'second', fallback=0)
 
     # 启用早安问好的群
-    group_id: List[int] = strtogroupid(DATA.get_config('morning', 'group_id'))
+    group_id: list[int] = strtogroupid(DATA.config.get('morning', 'group_id'))
 
     @validator('group_id', always=True)
-    def group_id_validator(cls, v: List[int]):
+    def group_id_validator(cls, v: list[int]):
         """ 验证并保存配置 """
-        DATA.set_config('morning', 'group_id', groupidtostr(v))
+        DATA.config.set('morning', 'group_id', groupidtostr(v))
         return v
 
     class Config:
