@@ -3,11 +3,12 @@
 https://github.com/Binaryify/NeteaseCloudMusicApi
 """
 from typing import Optional
+from nonebot.adapters.cqhttp import MessageSegment, Message
 
 import httpx
 
 
-async def call_netease_api(name: str) -> Optional[str]:
+async def call_netease_api(name: str) -> Optional[Message]:
     """ 网易云搜索 API """
     if not name:
         return None
@@ -27,7 +28,8 @@ async def call_netease_api(name: str) -> Optional[str]:
 
             # 获取音乐 ID，并返回对应的 CQ 码
             music_id = rjson['result']['songs'][0]['id']
-            return f'[CQ:music,type=163,id={music_id}]'
+
+            return Message(MessageSegment.music('163', music_id))
 
     except (httpx.HTTPError, KeyError):
         # 抛出上面任何异常，说明调用失败
