@@ -5,13 +5,13 @@ from pydantic import BaseSettings
 
 from src.utils.plugin import PluginData
 
-DATA = PluginData('robot', config=True)
+DATA = PluginData('robot')
 
 
 class Config(BaseSettings):
-    tencent_ai_app_id: str = DATA.get_config('tencent', 'app_id')
-    tencent_ai_app_key: str = DATA.get_config('tencent', 'app_key')
-    tuling_api_key: str = DATA.get_config('tuling', 'api_key')
+    tencent_ai_secret_id: str = DATA.config.get('tencent', 'secret_id')
+    tencent_ai_secret_key: str = DATA.config.get('tencent', 'secret_key')
+    tuling_api_key: str = DATA.config.get('tuling', 'api_key')
 
     class Config:
         extra = 'ignore'
